Remove eslint and start reorganizing folders
