+++ package.json
@@ -40,7 +40,6 @@
     ]
   },
   "devDependencies": {
-    "eslint": "^8.12.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-jsx-a11y": "^6.5.1",
