+++ src/backend/client/firestore.js
@@ -6,58 +6,7 @@ import Listing from './../../objects/listing';
 import Property from './../../objects/property';
 import Item from './../../objects/item';
 
-const listingConverter = {
-    toFirestore: (listing) => {
-            return {
-                listingTitle: listing.item.title,
-                seller: listing.seller,
-                buyer: listing.buyer,
-                dateBought: Timestamp.fromDate(listing.dateBought),
-                quantity: listing.quantity,
-                isPurchased: listing.isPurchased,
-                shippingCost: listing.shippingCost,
-                shippingFrom: listing.shippingFrom,
-                id: listing.id,
-                item: {
-                    title: listing.item.title,
-                    price: listing.item.price,
-                    description: listing.item.description,
-                    category: listing.item.category,
-                    keywords:listing.item.keywords,
-                    images: listing.item.images,
-                    physicalProperties: {
-                         height: listing.item.physicalProperties.height,
-                         width: listing.item.physicalProperties.width,
-                         depth: listing.item.physicalProperties.depth,
-                         weight: listing.item.physicalProperties.weight,
-                    }
-                }
-    
-            }
-    },
-    fromFirestore: (snapshot, options) => {
-        const data = snapshot.data(options);
-        return new Listing(
-            data.id,
-            data.seller,
-            data.buyer,
-            data.dateBought,
-            data.quantity,
-            data.isPurchased,
-            data.shippingCost,
-            data.shippingFrom,
-            new Item(
-                data.item.title,
-                data.item.price,
-                data.item.description,
-                data.item.category,
-                data.item.keywords,
-                data.item.images,
-                new Property(data.item.height, data.item.width, data.item.depth, data.item.weight)
-            )
-        );
-    }
-};
+
 
 const createListing = (listing, image) => {
     return new Promise((resolve, reject) => {
@@ -106,18 +55,6 @@ const getListings = async (qStr) => {
     
 }
 
-const getListing = async (listingId) => {
-    return new Promise(async (resolve, reject) => {
-                const listingRef = doc(db, "listings", listingId);
-                const listingSnap = await getDoc(listingRef);
-
-                if (listingSnap.exists()) {
-                    resolve(listingSnap.data());
-                } else {
-                    reject(new Error('Listing with this id does not exist!'));
-                }
-        });
-}
 
 const getAllListings = () => {
     return new Promise(async (resolve, reject) => {
