+++ src/objects/item.js
@@ -21,4 +21,8 @@ export class Item{
     }
 }
 
+
+
+
+
 export default Item;
